Updated bindings.

+++ examples/complex/bindings/csharp/Interop.cs
@@ -13,7 +13,7 @@ namespace My.Company
         public const string NativeLib = "example_complex";
 
         /// Call for a friend.
-        public const uint THE_MAGIC_CONSTANT = 666;
+        public const uint THE_MAGIC_CONSTANT = (uint) 666;
 
 
         /// Returns the version of this API.
